use [[]] style instead of if blocks, simplify config init file

+++ init/30_set_config.sh
@@ -1,16 +1,7 @@
 #!/bin/bash
+mkdir -p /config/dbase_and_logs /daapd-pidfolder
 
-if [ ! -d "/config/dbase_and_logs" ]; then
-mkdir -p /config/dbase_and_logs
-fi
-
-if [ ! -f "/config/forked-daapd.conf" ]; then
-cp /defaults/forked-daapd.conf /config/forked-daapd.conf
-fi
-
-if [ ! -d "/daapd-pidfolder" ]; then
-mkdir -p /daapd-pidfolder
-fi
+[[ ! -f "/config/forked-daapd.conf" ] && cp /defaults/forked-daapd.conf /config/forked-daapd.conf
 
 cp /config/forked-daapd.conf /etc/forked-daapd.conf
 chown -R abc:abc /config /daapd-pidfolder /app
